fix: resolve persistent lint errors and wrk2 arm64 build
- Fix lingering errcheck and unconvert lints in epoll/http2.go, epoll/hybrid.go, iouring/http2.go, iouring/hybrid.go
- Update userdata.tftpl to install libluajit-5.1-dev and build wrk2 with system LuaJIT to fix ARM64 build failure

+++ servers/theoretical/iouring/http2.go
@@ -161,7 +161,7 @@ func (s *HTTP2Server) setupBuffers() error {
 	sqe.OpcodeFlags = bufferGroup
 	sqe.UserData = 0
 
-	s.sqArray[idx] = uint32(idx)
+	s.sqArray[idx] = idx
 	*s.sqTail = tail + 1
 
 	_, _, errno := unix.Syscall6(unix.SYS_IO_URING_ENTER, uintptr(s.ringFd), 1, 0, 0, 0, 0)
